requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.0.3
gspread==5.10.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
Werkzeug==3.0.0